+++ CHANGELOG.rst
@@ -4,6 +4,14 @@ Changelog
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+[Unreleased]
+------------
+
+Added
+~~~~~
+* CLI flag for preventing pre-counting of files in folder
+
+
 [0.2.1] - 2016-05-11
 --------------------